src/pyproject.toml
[project]
name = "eegviz"
version = "0.1.0"
description = "EEG Visualizer using MNE-Python with a Streamlit UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mne==1.10.1",
  "numpy>=1.24",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "imageio>=2.31",
  "python-dateutil>=2.8",
  "streamlit>=1.33",
]

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.4", "black>=24", "mypy>=1.8"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B","SIM"]
ignore = ["E203","E501"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
